# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cotopia Streaming Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /songs:
    get:
      operationId: listSongs
      tags: [songs]
      summary: List songs
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Search query
        - in: query
          name: artistId
          schema:
            type: integer
        - in: query
          name: genre
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SongList"
    post:
      operationId: createSong
      tags: [songs]
      summary: Create a new song (artist/label)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SongInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
...
  /songs/trending:
    get:
      operationId: getTrendingSongs
      tags: [songs]
      summary: Get trending songs
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Song"
  /songs/{id}:
    get:
      operationId: getSong
      tags: [songs]
      summary: Get a song by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SongDetail"
        "404":
          description: Not found
    patch:
      operationId: updateSong
      tags: [songs]
      summary: Update a song
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SongUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
    delete:
      operationId: deleteSong
      tags: [songs]
      summary: Delete a song
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
  /songs/{id}/play:
    post:
      operationId: recordSongPlay
      tags: [songs]
      summary: Record a song play event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
  /songs/{id}/comments:
    get:
      operationId: getSongComments
      tags: [comments]
      summary: Get comments for a song
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createSongComment
      tags: [comments]
      summary: Add a comment to a song
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /songs/{id}/rate:
    post:
      operationId: rateSong
      tags: [songs]
      summary: Rate a song (1-5)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RatingInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingResult"

  # ── Videos ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Song:
      type: object
      required: [id, title, artistId, artistName, duration, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        artistId:
          type: integer
        artistName:
          type: string
        albumId:
          type: ["integer", "null"]
        albumName:
          type: ["string", "null"]
        genre:
          type: ["string", "null"]
        duration:
          type: integer
        coverUrl:
          type: ["string", "null"]
        streamUrl:
          type: ["string", "null"]
        playCount:
          type: integer
        avgRating:
          type: ["number", "null"]
        status:
          type: string
          enum: [draft, pending_review, approved, rejected, published]
        createdAt:
          type: string
    SongDetail:
      allOf:
        - $ref: "#/components/schemas/Song"
        - type: object
          properties:
            commentCount:
              type: integer
            userRating:
              type: ["integer", "null"]
            isFavorited:
              type: boolean
    SongInput:
      type: object
      required: [title, duration]
      properties:
        title:
          type: string
        albumId:
          type: integer
        genre:
          type: string
        duration:
          type: integer
        coverUrl:
          type: string
        streamUrl:
          type: string
        lyrics:
          type: string
    SongUpdate:
      type: object
      properties:
        title:
          type: string
        genre:
          type: string
        coverUrl:
          type: string
        streamUrl:
          type: string
        lyrics:
          type: string
    SongList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Song"
        total:
          type: integer

    # ── Videos ──────────────────────────────────────────────────────────────
...
    Comment:
      type: object
      required: [id, userId, username, content, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
        content:
          type: string
        parentId:
          type: ["integer", "null"]
        replies:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        createdAt:
          type: string
    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        parentId:
          type: integer

    # ── Ratings ──────────────────────────────────────────────────────────────
    RatingInput:
      type: object
      required: [rating]
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
    RatingResult:
      type: object
      required: [avgRating, ratingCount]
      properties:
        avgRating:
          type: number
        ratingCount:
          type: integer

    # ── Submissions ───────────────────────────────────────────────────────────
...